refactor(sweeps): move scheduler package under wandb.sdk.sweeps
Relocate `wandb.sdk.sweep_scheduler` to `wandb.sdk.sweeps.scheduler` and
make `wandb.sdk.sweeps` the source of truth for `RunState`, which used to
live in `wandb.sdk.launch.sweeps.scheduler`. The old location re-exports it
so existing imports keep working.

Also rename `RunEnriched` to `RunWithMetrics` and `Optimizer.next_n_runs`
to `ask_n_runs` to match the ask-tell naming.
